+++ src/cone/firebase/testing/firebase_admin/credentials.py
@@ -1,4 +1,4 @@
-class Certificate(object):
+class Certificate:
 
     def __init__(self, service_account_json):
         self.service_account_json = service_account_json